Properly deal with duplicate sealed property names.
In order to properly deal with duplicate sealed attribute names,
we need to actually ensure that one value is read for each time
any attribute name is encountered.

Since this also needs to be handled for the case where the traits
are sent by reference, we need to actually store the possible
duplicated attribute names on the traits object, and only when
serializing the traits ensure we do not generate duplicate attribute
names in the serialized data.

Refs #7.
Closes #9.